Refactored FeaturedProducts Into seperate component
