+++ src/pages/index.tsx
@@ -1,12 +1,6 @@
-import { HTML5Backend } from 'react-dnd-html5-backend';
-import { DndProvider } from 'react-dnd';
-import { Container } from '~/components/FeaturedProducts/Container';
 import type { InferGetStaticPropsType, GetStaticProps } from 'next';
 import { Product } from '~/types/Product';
-import { Button } from '~/components/ui/button';
-import { useAppDispatch, useAppSelector } from '~/hooks/redux';
-import { setIsEditingFeaturedProducts } from '~/redux/reducers/global/globalSlice';
-import { useToast } from '~/components/ui/use-toast';
+import FeaturedProducts from '~/components/FeaturedProducts/FeaturedProducts';
 
 export const getStaticProps: GetStaticProps<{ data: Product[] }> = () => {
   return {
@@ -70,49 +64,11 @@ export const getStaticProps: GetStaticProps<{ data: Product[] }> = () => {
 export default function Home({
   data,
 }: InferGetStaticPropsType<typeof getStaticProps>) {
-  const isEditingFeatured = useAppSelector(
-    state => state.globalSlice.isEditingFeaturedProducts,
-  );
-  const dispatch = useAppDispatch();
-  const { toast } = useToast();
   return (
     <div className="flex">
       <div className=" w-full">
         <p className="border-b pb-3 text-5xl font-medium">Dashboard</p>
-        <p className="pt-4 text-4xl">Featured Products</p>
-        <p>
-          These are the featured products that appear on the homepage, you can
-          re-arrange the order by dragging and dropping and it will update the
-          main website.
-        </p>
-        <DndProvider backend={HTML5Backend}>
-          <div className="py-8">
-            <Container data={data} />
-          </div>
-        </DndProvider>
-        {isEditingFeatured ? (
-          <Button
-            className="bg-green-600 hover:bg-green-500"
-            onClick={() => {
-              dispatch(setIsEditingFeaturedProducts(false));
-              toast({
-                title: 'Saved ✔',
-                variant: 'success',
-                description:
-                  'The order of the featured products has been updated',
-              });
-            }}>
-            Save Order
-          </Button>
-        ) : (
-          <Button
-            className="bg-orange-400 hover:bg-orange-300"
-            onClick={() => {
-              dispatch(setIsEditingFeaturedProducts(true));
-            }}>
-            Edit Order
-          </Button>
-        )}
+        <FeaturedProducts data={data} />
       </div>
     </div>
   );
